Conflicts:
	core/src/main/java/cosmos/SortingMetadata.java
	core/src/test/java/cosmos/results/integration/GroupByIntegrationTest.java
	pom.xml
	test/src/main/java/cosmos/mapred/MediawikiIngestJob.java
	test/src/main/java/cosmos/mapred/MediawikiQueries.java
	test/src/test/java/cosmos/results/integration/CosmosIntegrationTest.java

This syncs up the accumulo-1.4.4 branch to what master (1.5.0) was at.
Changed modules from using a classifier to tagging each module with a
"-accumulo1.4" suffix in the artifactId.

+++ core/src/main/java/cosmos/results/PagedQueryResult.java
@@ -41,7 +41,7 @@ public class PagedQueryResult<T> implements Results<List<T>> {
   
   public PagedQueryResult(CloseableIterable<T> results, Paging limits) {
     source = results;
-    pagedLimitedResults = Iterables.partition(Iterables.limit(results, limits.maxResults().intValue()), limits.pageSize());
+    pagedLimitedResults = Iterables.partition(Iterables.limit(results, limits.maxResults()), limits.pageSize());
   }
   
   @Override
